Change compression and bin location
Change compression format to lzo for 2x startup speed, as seen in this [blogpost](https://snapcraft.io/blog/snap-speed-improvements-with-new-compression-algorithm).
I've tested it myself and on my (slow) machine it improved the cold start time from 14 seconds to 7, which is kind of impressive.

Also change bin location in the snap because the build command complained about it and changed it every time. So this suppresses the warning message.